+++ src/Products/PloneMeeting/tests/testUtils.py
@@ -2,27 +2,12 @@
 #
 # File: testUtils.py
 #
-# Copyright (c) 2017 by Imio.be
-#
 # GNU General Public License (GPL)
 #
-# This program is free software; you can redistribute it and/or
-# modify it under the terms of the GNU General Public License
-# as published by the Free Software Foundation; either version 2
-# of the License, or (at your option) any later version.
-#
-# This program is distributed in the hope that it will be useful,
-# but WITHOUT ANY WARRANTY; without even the implied warranty of
-# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-# GNU General Public License for more details.
-#
-# You should have received a copy of the GNU General Public License
-# along with this program; if not, write to the Free Software
-# Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
-# 02110-1301, USA.
-#
 
+from DateTime import DateTime
 from Products.PloneMeeting.tests.PloneMeetingTestCase import PloneMeetingTestCase
+from Products.PloneMeeting.utils import set_field_from_ajax
 from Products.PloneMeeting.utils import validate_item_assembly_value
 
 
@@ -56,6 +41,45 @@ class testUtils(PloneMeetingTestCase):
         # is bypassed, this let's edit an old wrong value
         self.request.set('initial_edit', u'1')
         self.assertTrue(validate_item_assembly_value(ASSEMBLY_WRONG_VALUE))
+
+    def test_pm_Set_field_from_ajax(self):
+        """Work on AT and DX."""
+        cfg = self.meetingConfig
+        cfg.setItemAdviceStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceEditStates((self._stateMappingFor('itemcreated'), ))
+        cfg.setItemAdviceViewStates((self._stateMappingFor('itemcreated'), ))
+
+        # item
+        self.changeUser('pmCreator1')
+        item = self.create('MeetingItem')
+        item.setOptionalAdvisers((self.vendors_uid, ))
+        item._update_after_edit()
+        new_value = "<p>My item description.</p>"
+        self.assertEqual(item.Description(), "")
+        self.assertFalse(self.catalog(Description="my item description"))
+        set_field_from_ajax(item, 'description', new_value)
+        self.assertEqual(item.Description(), new_value)
+        self.assertEqual(self.catalog(Description="my item description")[0].UID, item.UID())
+        self.assertEqual(self.catalog(SearchableText="my item description")[0].UID, item.UID())
+
+        # meeting
+        self.changeUser('pmManager')
+        meeting = self.create('Meeting', date=DateTime("2020/11/30"))
+        new_value = "<p>My meeting notes.</p>"
+        self.assertEqual(meeting.getNotes(), "")
+        self.assertFalse(self.catalog(SearchableText="my meeting notes"))
+        set_field_from_ajax(meeting, 'notes', new_value)
+        self.assertEqual(meeting.getNotes(), new_value)
+        self.assertEqual(self.catalog(SearchableText="my meeting notes")[0].UID, meeting.UID())
+
+        # advice
+        self.changeUser('pmReviewer2')
+        advice = self.addAdvice(item, advice_comment=u"")
+        new_value = "<p>My advice comment.</p>"
+        self.assertEqual(advice.advice_comment.raw, u"")
+        self.assertFalse(self.catalog(SearchableText="my advice comment"))
+        set_field_from_ajax(advice, 'advice_comment', new_value)
+        self.assertEqual(advice.advice_comment.output, new_value)
 
 
 def test_suite():