+++ src/Products/PloneMeeting/utils.py
@@ -937,35 +937,38 @@ def get_dx_widget(obj, field_name, mode=DISPLAY_MODE):
     return widget
 
 
-def setFieldFromAjax(obj, fieldName, newValue, remember=True, tranform=True, reindex=True, unlock=True):
+def set_field_from_ajax(obj, field_name, new_value, remember=True, tranform=True, reindex=True, unlock=True):
     '''Sets on p_obj the content of a field whose name is p_fieldName and whose
        new value is p_fieldValue. This method is called by Ajax pages.'''
 
     if IDexterityContent.providedBy(obj):
-        setattr(obj, fieldName, richtextval(newValue))
+        setattr(obj, field_name, richtextval(new_value))
     else:
-        field = obj.getField(fieldName)
+        field = obj.getField(field_name)
         if remember:
             # Keep old value, we might need to historize it.
-            previousData = rememberPreviousData(obj, fieldName)
-            field.getMutator(obj)(newValue, content_type='text/html')
+            previousData = rememberPreviousData(obj, field_name)
+            field.getMutator(obj)(new_value, content_type='text/html')
             # Potentially store it in object history
             if previousData:
                 addDataChange(obj, previousData)
         else:
-            field.getMutator(obj)(newValue, content_type='text/html')
+            field.getMutator(obj)(new_value, content_type='text/html')
 
     if tranform:
         # Apply XHTML transforms when relevant
-        transformAllRichTextFields(obj, onlyField=fieldName)
+        transformAllRichTextFields(obj, onlyField=field_name)
 
     if reindex:
         # only reindex relevant indexes aka SearchableText + field specific index if exists
         index_names = api.portal.get_tool('portal_catalog').indexes()
         extra_idxs = ['SearchableText']
-        probable_index_name = 'get%s%s' % (fieldName[0].upper(), fieldName[1:])
-        if fieldName in index_names:
-            extra_idxs.append(fieldName)
+        if field_name == 'description':
+            probable_index_name = 'Description'
+        else:
+            probable_index_name = 'get%s%s' % (field_name[0].upper(), field_name[1:])
+        if field_name in index_names:
+            extra_idxs.append(field_name)
         if probable_index_name in index_names:
             extra_idxs.append(probable_index_name)
         notifyModifiedAndReindex(obj, extra_idxs=extra_idxs)
@@ -974,7 +977,7 @@ def setFieldFromAjax(obj, fieldName, newValue, remember=True, tranform=True, rei
         unlockAfterModification(obj, event={})
     # add a fingerpointing log message
     extras = 'object={0} field_name={1}'.format(
-        repr(obj), fieldName)
+        repr(obj), field_name)
     fplog('quickedit_field', extras=extras)